refactor: moved over Qurey Builder
